+++ library/src/main/java/io/agora/board/fast/FastException.java
@@ -1,20 +1,32 @@
 package io.agora.board.fast;
 
 import androidx.annotation.IntDef;
-
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 public class FastException extends RuntimeException {
+
     public static final int TYPE_SDK = 0;
+
     public static final int TYPE_ROOM = 1;
+
     public static final int TYPE_PLAYER = 2;
+
     public static final int SDK_ERROR = 100;
+
+    public static final int SDK_SETUP_ERROR = 101;
+
     public static final int ROOM_JOIN_ERROR = 200;
+
     public static final int ROOM_DISCONNECT_ERROR = 201;
+
+    public static final int ROOM_KICKED = 202;
+
     public static final int PLAYER_JOIN_ERROR = 200;
+
     private int type;
+
     private int code;
 
     private FastException(@Type int type, int code, String message) {
@@ -31,6 +43,14 @@ public class FastException extends RuntimeException {
 
     public static FastException createSdk(String message) {
         return new FastException(TYPE_SDK, SDK_ERROR, message);
+    }
+
+    public static FastException createSdk(int code, String message) {
+        return new FastException(TYPE_SDK, code, message);
+    }
+
+    public static FastException createRoom(int code, String message) {
+        return new FastException(TYPE_ROOM, code, message);
     }
 
     public static FastException createRoom(int code, String message, Throwable throwable) {
@@ -53,5 +73,6 @@ public class FastException extends RuntimeException {
     @Retention(RetentionPolicy.SOURCE)
     @IntDef({TYPE_SDK, TYPE_ROOM, TYPE_PLAYER})
     public @interface Type {
+
     }
 }