+++ library/src/main/java/io/agora/board/fast/FastRoom.java
@@ -2,6 +2,8 @@ package io.agora.board.fast;
 
 import static io.agora.board.fast.FastException.ROOM_DISCONNECT_ERROR;
 import static io.agora.board.fast.FastException.ROOM_JOIN_ERROR;
+import static io.agora.board.fast.FastException.ROOM_KICKED;
+import static io.agora.board.fast.FastException.SDK_SETUP_ERROR;
 
 import android.view.View;
 import android.view.ViewGroup;
@@ -29,11 +31,6 @@ import com.herewhite.sdk.domain.RoomState;
 import com.herewhite.sdk.domain.SDKError;
 import com.herewhite.sdk.domain.Scene;
 import com.herewhite.sdk.domain.WindowAppParam;
-
-import org.json.JSONObject;
-
-import java.util.UUID;
-
 import io.agora.board.fast.extension.ErrorHandler;
 import io.agora.board.fast.extension.FastResource;
 import io.agora.board.fast.extension.FastResult;
@@ -47,6 +44,8 @@ import io.agora.board.fast.internal.FastOverlayManager;
 import io.agora.board.fast.internal.FastRoomContext;
 import io.agora.board.fast.internal.FastRoomPhaseHandler;
 import io.agora.board.fast.internal.PromiseResultAdapter;
+import io.agora.board.fast.internal.Util;
+import io.agora.board.fast.internal.WhiteboardViewManager;
 import io.agora.board.fast.model.ConverterType;
 import io.agora.board.fast.model.DocPage;
 import io.agora.board.fast.model.FastAppliance;
@@ -59,6 +58,8 @@ import io.agora.board.fast.ui.FastRoomController;
 import io.agora.board.fast.ui.LoadingLayout;
 import io.agora.board.fast.ui.OverlayLayout;
 import io.agora.board.fast.ui.RoomControllerGroup;
+import java.util.UUID;
+import org.json.JSONObject;
 
 public class FastRoom {
 
@@ -79,7 +80,9 @@ public class FastRoom {
     private final CommonCallback commonCallback = new CommonCallback() {
         @Override
         public void throwError(Object args) {
-
+            String message = Util.toJson(args);
+            FastLogger.error("sdk throw error " + message);
+            fastRoomContext.notifyFastError(FastException.createSdk(message));
         }
 
         @Override
@@ -90,7 +93,7 @@ public class FastRoom {
         @Override
         public void sdkSetupFail(SDKError error) {
             FastLogger.error("sdk setup fail ", error);
-            fastRoomContext.notifyFastError(FastException.createSdk(error.getMessage()));
+            fastRoomContext.notifyFastError(FastException.createSdk(SDK_SETUP_ERROR, error.getMessage()));
         }
 
         @Override
@@ -118,11 +121,12 @@ public class FastRoom {
         @Override
         public void onKickedWithReason(String reason) {
             FastLogger.warn("receive kicked from js with reason " + reason);
+            fastRoomContext.notifyFastError(FastException.createRoom(ROOM_KICKED, reason));
         }
 
         @Override
         public void onRoomStateChanged(RoomState modifyState) {
-            notifyRoomState(modifyState);
+            fastRoomContext.notifyRoomStateChanged(modifyState);
         }
 
         @Override
@@ -148,7 +152,7 @@ public class FastRoom {
         public void then(Room room) {
             FastLogger.info("join room success" + room.toString());
             FastRoom.this.room = room;
-            notifyRoomState(room.getRoomState());
+            updateRoomState(room.getRoomState());
             updateWritable();
             updateIfTextAppliance();
             notifyRoomReady();
@@ -160,6 +164,28 @@ public class FastRoom {
             fastRoomContext.notifyFastError(FastException.createRoom(ROOM_JOIN_ERROR, t.getMessage(), t));
         }
     };
+
+    private void updateRoomState(RoomState roomState) {
+        if (roomState.getBroadcastState() != null) {
+            roomControllerGroup.updateBroadcastState(roomState.getBroadcastState());
+        }
+
+        if (roomState.getMemberState() != null) {
+            roomControllerGroup.updateMemberState(roomState.getMemberState());
+        }
+
+        if (roomState.getSceneState() != null) {
+            roomControllerGroup.updateSceneState(roomState.getSceneState());
+        }
+
+        if (roomState.getPageState() != null) {
+            roomControllerGroup.updatePageState(roomState.getPageState());
+        }
+
+        if (roomState.getWindowBoxState() != null) {
+            roomControllerGroup.updateWindowBoxState(roomState.getWindowBoxState());
+        }
+    }
 
     /**
      * workaround for text appliance when init state
@@ -183,25 +209,7 @@ public class FastRoom {
 
         @Override
         public void onRoomStateChanged(RoomState roomState) {
-            if (roomState.getBroadcastState() != null) {
-                roomControllerGroup.updateBroadcastState(roomState.getBroadcastState());
-            }
-
-            if (roomState.getMemberState() != null) {
-                roomControllerGroup.updateMemberState(roomState.getMemberState());
-            }
-
-            if (roomState.getSceneState() != null) {
-                roomControllerGroup.updateSceneState(roomState.getSceneState());
-            }
-
-            if (roomState.getPageState() != null) {
-                roomControllerGroup.updatePageState(roomState.getPageState());
-            }
-
-            if (roomState.getWindowBoxState() != null) {
-                roomControllerGroup.updateWindowBoxState(roomState.getWindowBoxState());
-            }
+            updateRoomState(roomState);
         }
 
         @Override
@@ -302,10 +310,6 @@ public class FastRoom {
         if (room.getWritable()) {
             room.disableSerialization(false);
         }
-    }
-
-    private void notifyRoomState(RoomState roomState) {
-        fastRoomContext.notifyRoomStateChanged(roomState);
     }
 
     private void notifyRoomReady() {
@@ -680,8 +684,8 @@ public class FastRoom {
 
     public void destroy() {
         WhiteboardView whiteboardView = fastboardView.whiteboardView;
-        whiteboardView.removeAllViews();
-        whiteboardView.destroy();
+        fastboardView.removeView(whiteboardView);
+        WhiteboardViewManager.get().release(whiteboardView);
     }
 
     public RoomControllerGroup getRootRoomController() {